feat: 프로필 페이지 리팩토링을 통한 코드 정리 및 기능 분리 (#162)
